Fixes #621 - Removed unnecessary bundling of entire node_modules for handling native modules
