+++ package.json
@@ -21,7 +21,7 @@
   "scripts": {
     "start": "electron-forge start",
     "postinstall": "electron-builder install-app-deps",
-    "publish": "electron-forge package && node packager.js",
+    "publish": "electron-forge package",
     "lint": "eslint --ext .ts,.tsx .",
     "test": "jest",
     "test:watch": "jest --watch",
@@ -33,6 +33,7 @@
     "@electron-forge/plugin-webpack": "^6.4.2",
     "@fortawesome/fontawesome-free": "^6.4.2",
     "@svgr/webpack": "^8.1.0",
+    "@timfish/forge-externals-plugin": "^0.2.1",
     "@types/c3": "^0.7.10",
     "@types/d3": "^7.4.2",
     "@types/diff-match-patch": "^1.0.35",
@@ -48,8 +49,7 @@
     "@vercel/webpack-asset-relocator-loader": "^1.7.3",
     "css-loader": "^6.8.1",
     "electron": "^28.1.0",
-    "electron-builder": "^24.6.4",
-    "electron-packager": "^17.1.2",
+    "electron-builder": "^24.9.1",
     "eslint": "^8.53.0",
     "eslint-plugin-css-import-order": "^1.1.0",
     "eslint-plugin-import": "^2.29.0",
